firstweb/requirements.txt
django-jazzmin==2.6.0
django-summernote
django-taggit
django-embed-video
setuptools
shortuuid
Pillow
mdx-mdx